+++ src/vdisplay/capture/__init__.py
@@ -13,6 +13,7 @@ from .observation import (
     screen_observation_v1_schema,
 )
 from .pixels import downscale_rgb_nearest, resolve_capture_scale, rgb_mostly_black
+from .portal import capture_portal_png
 from .portal_screencast import (
     PortalScreenCastSession,
     get_active_screencast,
@@ -22,6 +23,27 @@ from .portal_screencast import (
     stop_screencast_session,
 )
 from .providers.engine import capture_full_png, capture_region_png, list_capture_providers
+from .providers.observation import (
+    MonitorSpec,
+    ObservationBatch,
+    ObservationProvider,
+    ObservationProviderChainError,
+    ObservationProviderFailure,
+    ProviderAvailability,
+    capture_observations_with_fallback,
+    coerce_screen_observation,
+    screen_observation_from_png,
+)
+from .providers.observation_builtin import (
+    BlackFrameError,
+    CliToolsObservationProvider,
+    GrimObservationProvider,
+    MssObservationProvider,
+    PortalScreenCastObservationProvider,
+    PortalScreenshotObservationProvider,
+    command_candidates,
+    run_png_command,
+)
 from .screencast_crop import resolve_multi_stream_region
 
 __all__ = [
@@ -30,8 +52,21 @@ __all__ = [
     "capture_full_png",
     "capture_host_png",
     "capture_host_to_file",
+    "capture_portal_png",
     "capture_region_png",
     "downscale_rgb_nearest",
+    "BlackFrameError",
+    "CliToolsObservationProvider",
+    "GrimObservationProvider",
+    "MonitorSpec",
+    "MssObservationProvider",
+    "ObservationBatch",
+    "ObservationProvider",
+    "ObservationProviderChainError",
+    "ObservationProviderFailure",
+    "PortalScreenCastObservationProvider",
+    "PortalScreenshotObservationProvider",
+    "ProviderAvailability",
     "SCREEN_OBSERVATION_V1",
     "ScreenObservation",
     "ScreenObservationV1",
@@ -48,6 +83,11 @@ __all__ = [
     "resolve_capture_scale",
     "resolve_multi_stream_region",
     "screen_observation_v1_schema",
+    "screen_observation_from_png",
+    "capture_observations_with_fallback",
+    "coerce_screen_observation",
+    "command_candidates",
+    "run_png_command",
     "start_screencast_session",
     "stop_screencast_session",
     "rgb_mostly_black",